# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Threat Modeling Agent API
  version: 0.1.0
  description: |
    REST API for the Threat Modeling Agent SaaS platform.

    ## Authentication
    All endpoints (except the auth callback) require a valid Bearer token issued by WorkOS.
    Tokens are validated server-side on every request. Client-side validation is not sufficient.

    ## Tenant scoping
    All data is scoped to an organisation. The `org_id` is extracted from the validated JWT —
    it is never accepted from request parameters or body.

    ## Rate limiting
    Authentication and job submission endpoints are rate-limited.
    See individual endpoint descriptions for details.

    ## Security
    This API is governed by CLAUDE.md and docs/specs/06-security.md.
    All responses to authenticated requests include `Cache-Control: no-store`.
servers:
  - url: https://api.tmsa.example.com/v1
    description: Production (West Europe)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: WorkOS-issued JWT. Validated server-side on every request.
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Organisation UUID. Must match the org_id in the authenticated user's JWT.
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code. Never contains stack traces or internal paths.
        message:
          type: string
          description: Human-readable summary. Never contains internal details.
        correlationId:
          type: string
          format: uuid

    # ── Pagination ───────────────────────────────────────────────────────────────
...
    OrgSummary:
      type: object
      required: [id, name, slug, role]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        role:
          $ref: '#/components/schemas/OrgMemberRole'
    OrgDetail:
      type: object
      required: [id, name, slug, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        hasCustomIdp:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateOrgRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 255
        slug:
          type: string
          minLength: 2
          maxLength: 63
          pattern: '^[a-z0-9][a-z0-9\-]*[a-z0-9]$'
    UpdateOrgRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 255
    OrgMemberRole:
      type: string
      enum: [owner, member]

    # ── Members ──────────────────────────────────────────────────────────────────
...
paths:
...
  /orgs:
    get:
      operationId: listOrgs
      summary: List organisations the authenticated user belongs to
      tags: [organisations]
      responses:
        '200':
          description: List of organisations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrgSummary'

    post:
      operationId: createOrg
      summary: Create a new organisation. Caller becomes owner.
      tags: [organisations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrgRequest'
      responses:
        '201':
          description: Organisation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgDetail'
        '409':
          description: Slug already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orgs/{orgId}:
    parameters:
      - $ref: '#/components/parameters/OrgId'

    get:
      operationId: getOrg
      summary: Get organisation details
      tags: [organisations]
      responses:
        '200':
          description: Organisation detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgDetail'
        '403':
          description: Not a member of this organisation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      operationId: updateOrg
      summary: Update organisation. Requires org:owner.
      tags: [organisations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrgRequest'
      responses:
        '200':
          description: Updated organisation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgDetail'
        '403':
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: deleteOrg
      summary: Delete organisation and all its data. Requires org:owner. Irreversible.
      tags: [organisations]
      responses:
        '204':
          description: Organisation deleted
        '403':
          description: Insufficient role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Members ──────────────────────────────────────────────────────────────────
...